Make the API compatible with miekg/dns: msg.Pack() should accept no parameters
In cloudflare/dns we made msg.Pack(buf) to accept a slice of bytes as
parameter. The idea is bold (ie: dns library should be able to reuse
byte buffers) but making the API incompatible with miekg/dns is a bad
idea. This path rolls back the API change and introduces a separate
Pack variant that accepts a buffer:

    msg.PackWithBuffer([]byte)